feat(splash, boxes, box): refactor box and boxes into their own components, async load boxes
