+++ src/components/box/index.test.tsx
@@ -1,0 +1,11 @@
+// import React from 'react';
+// import { render, unmountComponentAtNode } from 'react-dom';
+// import Box from '.';
+
+it('renders without crashing', () => {
+  // const div = document.createElement('div');
+  // render(<Box />, div);
+  // unmountComponentAtNode(div);
+
+  expect(true).toBe(true);
+});
