+++ src/components/boxes/index.test.tsx
@@ -1,0 +1,11 @@
+// import React from 'react';
+// import { render, unmountComponentAtNode } from 'react-dom';
+// import Boxes from '.';
+
+it('renders without crashing', () => {
+  // const div = document.createElement('div');
+  // render(<Boxes />, div);
+  // unmountComponentAtNode(div);
+
+  expect(true).toBe(true);
+});
